Split up test_home.py into urls and templates and moved test_protocols into views/models.
I removed the count query checks because it wasn't testing the views, templates, or URLs.  It was only checking the counts returned by the queries, which is not what the view tests (or templates or urls) are for.

Files checked in: DataRepo/tests/views/test_home.py DataRepo/tests/views/test_protocol.py DataRepo/tests/templates/test_home.py DataRepo/tests/test_urls.py DataRepo/tests/views/models/test_protocol.py
