+++ DataRepo/tests/views/models/test_protocol.py
@@ -2,27 +2,7 @@ from django.core.management import call_command
 from django.urls import reverse
 
 from DataRepo.models import Protocol
-from DataRepo.models.maintained_model import (
-    MaintainedModel,
-    UncleanBufferError,
-)
 from DataRepo.tests.tracebase_test_case import TracebaseTestCase
-
-
-def assert_coordinator_state_is_initialized():
-    # Obtain all coordinators that exist
-    all_coordinators = [MaintainedModel._get_default_coordinator()]
-    all_coordinators.extend(MaintainedModel._get_coordinator_stack())
-    if 1 != len(all_coordinators):
-        raise ValueError(
-            f"Before setting up test data, there are {len(all_coordinators)} MaintainedModelCoordinators."
-        )
-    if all_coordinators[0].auto_update_mode != "always":
-        raise ValueError(
-            "Before setting up test data, the default coordinator is not in always autoupdate mode."
-        )
-    if 0 != all_coordinators[0].buffer_size():
-        raise UncleanBufferError()
 
 
 class ProtocolViewTests(TracebaseTestCase):
@@ -34,9 +14,6 @@ class ProtocolViewTests(TracebaseTestCase):
 
     @classmethod
     def setUpTestData(cls):
-        # Ensure the auto-update buffer is empty.  If it's not, then a previously run test didn't clean up after itself
-        assert_coordinator_state_is_initialized()
-
         call_command("loaddata", "lc_methods")
         call_command("legacy_load_study", "DataRepo/data/tests/dataframes/loading.yaml")
